Interface changes

+++ My Project/src/Interface/B.java
@@ -3,7 +3,7 @@ package Interface;
 public interface B 
 {
 	public void sum();
-	public double Add();
+	public int Add();
 }
